Disable automatic allocation on credit note from invoice.
When creating a credit from an invoice, FA tries to allocate it to the
original invoice. It does so my unallocating payments until there is
enough to allocate the credit note itself. This however doesn't work
if a credit note has already been allocated instead of a payment.
Problems appear if you unallocate the credit note and try to resubmit
the new credit note. For some reason, resubmitting the credit note
updates the new credit note (which doesn't exist yet) which result
in modifying the original invoice instead of the credit note.

This automatic allocation is inconvenient anyway if the customer pays
the invoice in full but after a credit note has been raised.
It's therefore much easier to disable this feature and allocate the
credit note manually.